refactor/Minor refactor
