+++ packages/nocodb/src/__tests__/unit/rest/init/cleanupMeta.ts
@@ -1,4 +1,3 @@
-import Base from '../../../../lib/models/Base';
 import Model from '../../../../lib/models/Model';
 import Project from '../../../../lib/models/Project';
 import NcConnectionMgrv2 from '../../../../lib/utils/common/NcConnectionMgrv2';
